move cloneable body to node server
