+++ packages/next/server/base-server.ts
@@ -16,7 +16,6 @@ import type { ParsedUrlQuery } from 'querystring'
 import type { RenderOpts, RenderOptsPartial } from './render'
 import type { ResponseCacheEntry, ResponseCacheValue } from './response-cache'
 import type { UrlWithParsedQuery } from 'url'
-import type { TLSSocket } from 'tls'
 import {
   CacheFs,
   NormalizeError,
@@ -27,7 +26,6 @@ import {
 import type { PreviewData } from 'next/types'
 import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
 import type { BaseNextRequest, BaseNextResponse } from './base-http'
-import type { NodeNextRequest } from './base-http/node'
 import type { PayloadOptions } from './send-payload'
 
 import { join, resolve } from '../shared/lib/isomorphic/path'
@@ -80,7 +78,6 @@ import { getHostname } from '../shared/lib/get-hostname'
 import { parseUrl as parseUrlUtil } from '../shared/lib/router/utils/parse-url'
 import { getNextPathnameInfo } from '../shared/lib/router/utils/get-next-pathname-info'
 import { normalizePagePath } from '../shared/lib/page-path/normalize-page-path'
-import { getClonableBody } from './body-streams'
 
 export type FindComponentsResult = {
   components: LoadComponentsReturnType
@@ -230,6 +227,10 @@ export default abstract class Server<ServerOptions extends Options = Options> {
   protected abstract getRoutesManifest(): CustomRoutes
   protected abstract getPrerenderManifest(): PrerenderManifest
   protected abstract getServerComponentManifest(): any
+  protected abstract attachRequestMeta(
+    req: BaseNextRequest,
+    parsedUrl: NextUrlWithParsedQuery
+  ): void
 
   protected abstract sendRenderResult(
     req: BaseNextRequest,
@@ -419,22 +420,7 @@ export default abstract class Server<ServerOptions extends Options = Options> {
         parsedUrl.query = parseQs(parsedUrl.query)
       }
 
-      const protocol = (
-        (req as NodeNextRequest).originalRequest?.socket as TLSSocket
-      )?.encrypted
-        ? 'https'
-        : 'http'
-
-      // When there are hostname and port we build an absolute URL
-      const initUrl =
-        this.hostname && this.port
-          ? `${protocol}://${this.hostname}:${this.port}${req.url}`
-          : req.url
-
-      addRequestMeta(req, '__NEXT_INIT_URL', initUrl)
-      addRequestMeta(req, '__NEXT_INIT_QUERY', { ...parsedUrl.query })
-      addRequestMeta(req, '_protocol', protocol)
-      addRequestMeta(req, '__NEXT_CLONABLE_BODY', getClonableBody(req.body))
+      this.attachRequestMeta(req, parsedUrl)
 
       const domainLocale = detectDomainLocale(
         this.nextConfig.i18n?.domains,
